pyproject.toml
[project]
name = "fuzzy"
version = "7.0.0"
description = "The iron-booted robotic resident of The Tsunami Zone. "
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hikari[speedups]>=2.3.3,<3.0",
    "hikari-crescent>=1.3,<2.0",
    "python-dotenv>=1.1,<2.0",
    "uvloop==0.21.*",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
